requirements.txt
scapy==2.5.0
aiohttp==3.9.1
asyncio-throttle==1.0.2
click==8.1.7
colorama==0.4.6
tqdm==4.66.1
pytest==7.4.3
pytest-asyncio==0.21.1
pyyaml==6.0.1